Move progress under deck
